Begin MVC architecture
