+++ src/main/resources/static/old/example.html
@@ -14,6 +14,6 @@
 		<script src="https://cdnjs.cloudflare.com/ajax/libs/cytoscape/3.18.2/cytoscape.min.js"></script>
 		<!--<script src="https://unpkg.com/webcola/WebCola/cola.min.js"></script>-->
 		<!--<script src="https://cdn.jsdelivr.net/npm/cytoscape-cola@2.2.3/cytoscape-cola.min.js"></script>-->
-		<script src="../main.js" type="module"></script>
+		<script src="../app.js" type="module"></script>
 	</body>
 </html>
