+++ src/main/resources/static/src/controller/apiController.js
@@ -1,0 +1,12 @@
+/**
+ * Controller for communicating with the API.
+ */
+class ApiController {
+
+    // TODO: finish this
+    constructor(model) {
+        this.model = model;
+    }
+}
+
+export default ApiController;
