+++ src/main/resources/static/src/model/systemModel.js
@@ -1,0 +1,15 @@
+/**
+ * Model for the application.
+ */
+class SystemModel {
+
+    /**
+     * Creates a new instance.
+     * @param systemOverview An instance of SystemOverview
+     */
+    constructor(systemOverview) {
+        this.systemOverview = systemOverview;
+    }
+}
+
+export default SystemModel;
